+++ src/apps/csv-import.cpp
@@ -39,6 +39,7 @@ int main(int argc, char **argv) try
     transit::gtfs::CSVDiscSource source(path.c_str());
 
     auto dataset = transit::gtfs::readCSV(source);
+    dataset.filter_unreachable_stops();
     auto const message = transit::adaptor::Dictionary::encode(dataset.dictionary);
     transit::tool::container::StringTable dictionary;
     transit::adaptor::Dictionary::decode_into(dictionary, message);