+++ src/service/master.cpp
@@ -28,6 +28,7 @@ Master::Master(std::string const &path)
 {
     // the constructor only reads the base data. All other classes / look-ups are constructed on
     // demand
+    base_data.filter_unreachable_stops();
 }
 
 // request the plugin from the master service