+++ Changelog.md
@@ -9,3 +9,4 @@
   - added node bindings to expose plugin architecture
   - added basic frontend for debug tiles
   - added look-ups (plain, computational expensive) for coordinate to stop, stop to trip, stop to station
+  - implemented filtering unused stops to avoid look-up pollution